refactoring react components structure
