+++ Frontend/financial-advisor/src/components/About/About.jsx
@@ -1,4 +1,4 @@
-import about from "../assets/about.jpg"
+import about from "../../assets/about.jpg"
 import React from "react";
