+++ Frontend/financial-advisor/src/components/Footer/Footer.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './Footer.css'; 
-import logo from '../assets/logo.png'; 
+import logo from '../../assets/logo.png'; 
 
 function Footer() {
   return (
